More Input into its own module
