+++ compiler/src/fuzzer/utils.rs
@@ -11,34 +11,6 @@ use crate::{
     },
 };
 
-#[derive(Clone)]
-pub struct Input {
-    pub heap: Heap,
-    pub arguments: Vec<Pointer>,
-}
-impl fmt::Display for Input {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            self.arguments
-                .iter()
-                .map(|arg| arg.format(&self.heap))
-                .join(" "),
-        )
-    }
-}
-impl Input {
-    pub fn clone_to_other_heap(&self, other: &mut Heap) -> Vec<Pointer> {
-        self.heap
-            .clone_multiple_to_other_heap_with_existing_mapping(
-                other,
-                &self.arguments,
-                &mut FxHashMap::default(),
-            )
-    }
-}
-
 pub fn collect_symbols_in_heap(heap: &Heap) -> Vec<String> {
     heap.all_objects()
         .filter_map(|object| {
